+++ src/main/java/com/evan/blog/service/ArticleService.java
@@ -25,6 +25,6 @@ public interface ArticleService {
 
     void addArticle(TempArticle tempArticle);
 
-    void deleteArticles(List<Integer> pubIds) throws Exception;
+    void deleteArticles(List<Long> pubIds) throws Exception;
 
 }
